+++ src/zc/relationship/tests.py
@@ -48,7 +48,7 @@ class Demo(persistent.Persistent):
         self.id = id
 
     def __repr__(self):
-        return '<{} {}>'.format(self.__class__.__name__, self.id)
+        return f'<{self.__class__.__name__} {self.id}>'
 
 
 def keyrefSetUp(test):