+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changes
 
 - Drop support for Python 2.7, 3.5, 3.6.
 
+- Drop support for not existing ``zc.listcontainer``.
 
 2.1 (2021-03-22)
 ================